fix: moving some files out of preview code
